Edge server code changed eucDist from variable type from int to double

+++ SocketCode/EdgeServer.cpp
@@ -283,8 +283,8 @@ int main(int argc, char const *argv[]) {
     for(int i = 0; i < NUM_NODES; ++i) {
     	int matchIndex = -1;
 			int tmpMatchIndex = -1;
-			int tmpEucDist = 1000;
-			int minEucDist = 1000;
+			double tmpEucDist = 1000;
+			double minEucDist = 1000;
 			int max = -1;
 			int useIndex = -1;
 			bool updateFlag = false;
